Refactor configuration properties to separate files
- Extract inner classes (JhipsterCmd, NpmCmd, Github, Gitlab, Mail) from ApplicationProperties into dedicated configuration property classes:
    - JhipsterCmdProperties
    - NpmCmdProperties
    - GithubProperties
    - GitlabProperties
    - MailProperties
- Update ApplicationProperties to use these new bean types
- Improves modularity and adheres to the Single Responsibility Principle, making configuration management easier to maintain and test
