+++ src/main/java/io/github/jhipster/online/config/ApplicationProperties.java
@@ -20,40 +20,34 @@ package io.github.jhipster.online.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-/**
- * Properties specific to Jhonline.
- * <p>
- * Properties are configured in the application.yml file.
- * See {@link io.github.jhipster.config.JHipsterProperties} for a good example.
- */
 @ConfigurationProperties(prefix = "application", ignoreUnknownFields = false)
 public class ApplicationProperties {
 
-    private JhipsterCmd jhipsterCmd = new JhipsterCmd();
+    private JhipsterCmdProperties jhipsterCmd = new JhipsterCmdProperties();
 
-    private NpmCmd npmCmd = new NpmCmd();
+    private NpmCmdProperties npmCmd = new NpmCmdProperties();
 
-    private final Github github = new Github();
+    private final GithubProperties github = new GithubProperties();
 
-    private final Gitlab gitlab = new Gitlab();
+    private final GitlabProperties gitlab = new GitlabProperties();
 
-    private final Mail mail = new Mail();
+    private final MailProperties mail = new MailProperties();
 
     private String tmpFolder = System.getProperty("java.io.tmpdir");
 
-    public JhipsterCmd getJhipsterCmd() {
+    public JhipsterCmdProperties getJhipsterCmd() {
         return jhipsterCmd;
     }
 
-    public NpmCmd getNpmCmd() {
+    public NpmCmdProperties getNpmCmd() {
         return npmCmd;
     }
 
-    public void setNpmCmd(NpmCmd npmCmd) {
+    public void setNpmCmd(NpmCmdProperties npmCmd) {
         this.npmCmd = npmCmd;
     }
 
-    public void setJhipsterCmd(JhipsterCmd jhipsterCmd) {
+    public void setJhipsterCmd(JhipsterCmdProperties jhipsterCmd) {
         this.jhipsterCmd = jhipsterCmd;
     }
 
@@ -65,143 +59,15 @@ public class ApplicationProperties {
         this.tmpFolder = tmpFolder;
     }
 
-    public Github getGithub() {
+    public GithubProperties getGithub() {
         return github;
     }
 
-    public Gitlab getGitlab() {
+    public GitlabProperties getGitlab() {
         return gitlab;
     }
 
-    public Mail getMail() {
+    public MailProperties getMail() {
         return mail;
-    }
-
-    public static class JhipsterCmd {
-
-        private String cmd = "jhipster";
-        private Integer timeout = 120;
-
-        public String getCmd() {
-            return cmd;
-        }
-
-        public void setCmd(String cmd) {
-            this.cmd = cmd;
-        }
-
-        public Integer getTimeout() {
-            return timeout;
-        }
-
-        public void setTimeout(Integer timeout) {
-            this.timeout = timeout;
-        }
-    }
-
-    public static class NpmCmd {
-
-        private String cmd = "npm";
-        private Integer timeout = 120;
-
-        public String getCmd() {
-            return cmd;
-        }
-
-        public void setCmd(String cmd) {
-            this.cmd = cmd;
-        }
-
-        public Integer getTimeout() {
-            return timeout;
-        }
-
-        public void setTimeout(Integer timeout) {
-            this.timeout = timeout;
-        }
-    }
-
-    public static class Github {
-
-        private String clientId;
-        private String clientSecret;
-        private String host = "https://github.com";
-
-        public String getClientId() {
-            return clientId;
-        }
-
-        public void setClientId(String clientId) {
-            this.clientId = clientId;
-        }
-
-        public String getClientSecret() {
-            return clientSecret;
-        }
-
-        public void setClientSecret(String clientSecret) {
-            this.clientSecret = clientSecret;
-        }
-
-        public String getHost() {
-            return host;
-        }
-
-        public void setHost(String host) {
-            this.host = host;
-        }
-    }
-
-    public static class Gitlab {
-
-        private String clientId;
-        private String clientSecret;
-        private String host = "https://gitlab.com";
-        private String redirectUri;
-
-        public String getClientId() {
-            return clientId;
-        }
-
-        public void setClientId(String clientId) {
-            this.clientId = clientId;
-        }
-
-        public String getClientSecret() {
-            return clientSecret;
-        }
-
-        public void setClientSecret(String clientSecret) {
-            this.clientSecret = clientSecret;
-        }
-
-        public String getHost() {
-            return host;
-        }
-
-        public void setHost(String host) {
-            this.host = host;
-        }
-
-        public String getRedirectUri() {
-            return redirectUri;
-        }
-
-        public void setRedirectUri(String redirectUri) {
-            this.redirectUri = redirectUri;
-        }
-    }
-
-    public static class Mail {
-
-        private boolean enable;
-
-        public void setEnable(boolean enable) {
-            this.enable = enable;
-        }
-
-        public boolean isEnable() {
-            return enable;
-        }
     }
 }
