+++ src/main/java/io/github/jhipster/online/config/GithubProperties.java
@@ -1,0 +1,40 @@
+package io.github.jhipster.online.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration properties for GitHub integration.
+ */
+@Configuration
+@ConfigurationProperties(prefix = "application.github")
+public class GithubProperties {
+
+    private String clientId;
+    private String clientSecret;
+    private String host = "https://github.com";
+
+    public String getClientId() {
+        return clientId;
+    }
+
+    public void setClientId(String clientId) {
+        this.clientId = clientId;
+    }
+
+    public String getClientSecret() {
+        return clientSecret;
+    }
+
+    public void setClientSecret(String clientSecret) {
+        this.clientSecret = clientSecret;
+    }
+
+    public String getHost() {
+        return host;
+    }
+
+    public void setHost(String host) {
+        this.host = host;
+    }
+}
