+++ src/main/java/io/github/jhipster/online/config/GitlabProperties.java
@@ -1,0 +1,49 @@
+package io.github.jhipster.online.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration properties for GitLab integration.
+ */
+@Configuration
+@ConfigurationProperties(prefix = "application.gitlab")
+public class GitlabProperties {
+
+    private String clientId;
+    private String clientSecret;
+    private String host = "https://gitlab.com";
+    private String redirectUri;
+
+    public String getClientId() {
+        return clientId;
+    }
+
+    public void setClientId(String clientId) {
+        this.clientId = clientId;
+    }
+
+    public String getClientSecret() {
+        return clientSecret;
+    }
+
+    public void setClientSecret(String clientSecret) {
+        this.clientSecret = clientSecret;
+    }
+
+    public String getHost() {
+        return host;
+    }
+
+    public void setHost(String host) {
+        this.host = host;
+    }
+
+    public String getRedirectUri() {
+        return redirectUri;
+    }
+
+    public void setRedirectUri(String redirectUri) {
+        this.redirectUri = redirectUri;
+    }
+}
